test: PathAcceptanceTest 성공
